kuring-69 [수정] 유저 로직과 테스트 코드를 :data:user로 옮김
