+++ app/src/main/java/com/ku_stacks/ku_ring/di/RepositoryModule.kt
@@ -2,7 +2,6 @@ package com.ku_stacks.ku_ring.di
 
 import com.ku_stacks.ku_ring.data.api.DepartmentClient
 import com.ku_stacks.ku_ring.data.api.SendbirdClient
-import com.ku_stacks.ku_ring.data.db.BlackUserDao
 import com.ku_stacks.ku_ring.data.db.DepartmentDao
 import com.ku_stacks.ku_ring.preferences.PreferenceUtil
 import com.ku_stacks.ku_ring.repository.*
@@ -31,14 +30,6 @@ object RepositoryModule {
         sendbirdClient: SendbirdClient
     ): SendbirdRepository {
         return SendbirdRepositoryImpl(sendbirdClient)
-    }
-
-    @Provides
-    @Singleton
-    fun provideUserRepository(
-        blackUserDao: BlackUserDao
-    ): UserRepository {
-        return UserRepositoryImpl(blackUserDao)
     }
 
     @Provides
