+++ data/user/src/main/java/com/ku_stacks/ku_ring/user/repository/UserRepositoryImpl.kt
@@ -1,7 +1,7 @@
-package com.ku_stacks.ku_ring.repository
+package com.ku_stacks.ku_ring.user.repository
 
-import com.ku_stacks.ku_ring.data.db.BlackUserDao
-import com.ku_stacks.ku_ring.data.db.BlackUserEntity
+import com.ku_stacks.ku_ring.user.local.BlackUserDao
+import com.ku_stacks.ku_ring.user.local.BlackUserEntity
 import io.reactivex.rxjava3.core.Completable
 import io.reactivex.rxjava3.core.Single
 import javax.inject.Inject
