+++ data/user/src/test/resources/api-response/DefaultResponse.json
@@ -1,0 +1,5 @@
+{
+  "code": 200,
+  "message": "성공",
+  "data": null
+}
